refactor: 白饭相关 Host 代码抽离到 lib/rice-host.js
- 新增 lib/rice-host.js：RICE_CANDIDATES、riceLevelFromUrl、platformHeaders、loadRice、fetchAlertConfig（含 60s 缓存与令牌失效处理）
- lib/index.js 从 ./rice-host.js import 所需函数，删除 114 行内嵌重复实现
