+++ lib/index.js
@@ -2,6 +2,7 @@ import fs from 'node:fs'
 import os from 'node:os'
 import path from 'node:path'
 import { fileURLToPath } from 'node:url'
+import { riceLevelFromUrl, loadRice, fetchAlertConfig } from './rice-host.js'
 
 // Package root: lib/index.js -> package root. Keeps the bundle relocatable
 // when installed as a normal DSH npm plugin (node_modules or a local link).
@@ -55,42 +56,13 @@ function soundSetFromUrl(url) {
     return m ? decodeURIComponent(m[1]) : ''
   } catch (err) { return '' }
 }
-// 白饭档位解析：?level=full|half|empty，非法或缺省回退 full
-function riceLevelFromUrl(url) {
-  try {
-    const q = String(url || '').split('?')[1] || ''
-    const m = /(?:^|&)level=([^&]+)/.exec(q)
-    const v = m ? decodeURIComponent(m[1]) : ''
-    return v === 'half' || v === 'empty' ? v : 'full'
-  } catch (err) { return 'full' }
-}
-// 平台请求头（与官网前端一致的客户端标识，可帮助平台正确识别请求来源）
-function platformHeaders(token) {
-  return {
-    'Content-Type': 'application/json',
-    'Authorization': 'Bearer ' + token,
-    'x-client-bundle-id': 'com.deepseek.chat',
-    'x-client-locale': 'zh_CN',
-    'x-client-platform': 'web',
-    'x-client-version': '1.0.0',
-    'x-client-timezone-offset': String(-new Date().getTimezoneOffset() * 60),
-  }
-}
 const BALANCE_URL = 'https://api.deepseek.com/user/balance'
 const BALANCE_TTL_MS = 25000
-// 平台侧接口：users/current 可读取账户余额预警设置（balance_alert[币种].alert_bound）
-const PLATFORM_API = 'https://platform.deepseek.com'
 const RUA_GIF_CANDIDATES = [
   path.join(PACKAGE_ROOT, 'assets', 'rua.gif'),
   'D:/TestBox/deepseek/skin/rua.gif',
   'D:/TestBox/deepseek/rua.gif',
 ]
-// 白饭图标素材 — 三态：满碗 / 半碗 / 空碗（Issue #34）
-const RICE_CANDIDATES = {
-  full: [path.join(PACKAGE_ROOT, 'assets', 'rice', 'full.png')],
-  half: [path.join(PACKAGE_ROOT, 'assets', 'rice', 'half.png')],
-  empty: [path.join(PACKAGE_ROOT, 'assets', 'rice', 'empty.png')],
-}
 // DeepSeek CNY prices per million tokens: [空闲时段价, 高峰时段价].
 // 高峰时段：工作日 9:00–12:00 和 14:00–18:00（北京时间）；2026-08-23 起周末全天谷价。
 // Adjust here if DeepSeek changes pricing.
@@ -1690,24 +1662,6 @@ function apply(ctx) {
       throw new Error('whale image not found')
     }
 
-    // 白饭图标按档位读取（full/half/empty），带内存缓存
-    const riceBytesCache = {}
-    function loadRice(level) {
-      const key = level === 'half' || level === 'empty' ? level : 'full'
-      if (riceBytesCache[key]) return riceBytesCache[key]
-      const candidates = RICE_CANDIDATES[key] || RICE_CANDIDATES.full
-      for (const p of candidates) {
-        try {
-          const bytes = fs.readFileSync(p)
-          if (bytes && bytes.length > 0) {
-            riceBytesCache[key] = bytes
-            return bytes
-          }
-        } catch (err) {}
-      }
-      throw new Error('rice image not found: ' + key)
-    }
-
     function pickBalanceInfo(infos) {
       if (!Array.isArray(infos) || infos.length === 0) return null
       const num = (x) => (x && x.total_balance !== undefined ? Number(x.total_balance) : NaN)
@@ -1831,72 +1785,6 @@ function apply(ctx) {
         }
       }
       return found ? { amount: cost, tokens: tokens } : null
-    }
-
-    // —— 平台余额预警阈值（Issue #34 白饭图标档位参照）——
-    // 读取 DeepSeek 平台 users/current 的 balance_alert[币种]（enabled 开关 + alert_bound 阈值），
-    // 作为白饭图标「余额底线」的推荐值；无令牌/接口失败时前端回退菜单手动阈值。
-    let alertCache = null
-    function isAlertAuthFail(msg) {
-      return /authorization failed|invalid token|unauthorized|authentication failed/i.test(String(msg || ''))
-    }
-    async function fetchAlertConfig() {
-      const now = Date.now()
-      if (alertCache && now - alertCache.at < 60000) return alertCache.payload
-      let cred
-      try {
-        cred = await ctx.credentials.resolve('DEEPSEEK_PLATFORM_TOKEN')
-      } catch (err) {
-        return { ok: false, code: 'NO_TOKEN', error: '平台令牌读取失败' }
-      }
-      if (!cred) return { ok: false, code: 'NO_TOKEN', error: '未配置 DEEPSEEK_PLATFORM_TOKEN' }
-      const token = String(cred.value).replace(/^Bearer\s+/i, '')
-      let res
-      try {
-        res = await fetch(PLATFORM_API + '/auth-api/v0/users/current', {
-          headers: platformHeaders(token),
-          signal: AbortSignal.timeout(15000),
-        })
-      } catch (err) {
-        return { ok: false, code: 'NET', transient: true, error: String((err && err.message) || err).slice(0, 120) }
-      }
-      if (!res.ok) {
-        if (res.status === 401 || res.status === 403) {
-          alertCache = null
-          return { ok: false, code: 'TOKEN_EXPIRED', error: '令牌过期，请重新获取' }
-        }
-        return { ok: false, code: 'HTTP', error: '平台接口 HTTP ' + res.status }
-      }
-      let data
-      try {
-        data = await res.json()
-      } catch (err) {
-        return { ok: false, code: 'PARSE', error: '平台接口返回异常' }
-      }
-      const bizData = data && data.data && data.data.biz_data
-      if (!bizData || typeof bizData.currency !== 'string') {
-        return { ok: false, code: 'SHAPE', error: 'users/current 返回结构异常' }
-      }
-      const ba = bizData.balance_alert
-      const currency = bizData.currency.toUpperCase()
-      // 按账户币种取 balance_alert[currency]；该币种无条目则回退任意一个可用币种
-      let alert = ba && ba[currency]
-      if (!alert) {
-        const keys = Object.keys(ba || {})
-        const first = keys.find((k) => ba[k] && typeof ba[k].enabled === 'boolean' && 'alert_bound' in ba[k])
-        alert = first ? ba[first] : null
-      }
-      if (!alert || typeof alert.enabled !== 'boolean' || !('alert_bound' in alert)) {
-        return { ok: false, code: 'SHAPE', error: 'users/current 返回结构异常' }
-      }
-      const bound = Number(alert.alert_bound)
-      const payload = {
-        ok: true,
-        enabled: !!alert.enabled,
-        alertBound: isFinite(bound) ? bound : null,
-      }
-      alertCache = { at: now, payload }
-      return payload
     }
 
     function todayKey() {
@@ -2203,7 +2091,7 @@ function apply(ctx) {
       path: '/dsh-whale/alert.json',
       handler: async (req, res) => {
         try {
-          const payload = await fetchAlertConfig()
+          const payload = await fetchAlertConfig(ctx)
           res.writeHead(200, JSON_HEADERS)
           res.end(JSON.stringify(payload))
         } catch (err) {
